Separate Docker image and executor
